Read access to user account preference is added; updated account registration process such that Athlete role is created for each newly created account, and athlete is the default role for each account.

+++ util/type.go
@@ -3,7 +3,7 @@ package util
 func InterfaceSliceToIntSlice(input []interface{}) []int {
 	output := make([]int, 0)
 	for _, each := range input {
-		output = append(output, each.(int))
+		output = append(output, int(each.(float64)))
 	}
 	return output
 }
